+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Issue [#446](https://github.com/m-marini/wheellyj/issues/446): Change state machine configuration yaml
 - Issue [#455](https://github.com/m-marini/wheellyj/issues/455): Change radar map cell state update
+- Issue [#457](https://github.com/m-marini/wheellyj/issues/457):  Add rotation to target direction at move state
 
 ### Fixed